update(security): Ubuntu to alpine, update go pkgs (#6)
* update(security): Ubuntu to alpine, update go pkgs

The Ubuntu image had some vulnerablities and the easiest way to fix it was to switch to alpine. This results in a smaller image with fewer attack surfaces. Since this is a small image with a single application it makes sense to use a small base.

I've also updated some of the go packages, see the diff.

* update(security): Ubuntu to alpine, update go pkgs

The Ubuntu image had some vulnerablities and the easiest way to fix it was to switch to alpine. This results in a smaller image with fewer attack surfaces. Since this is a small image with a single application it makes sense to use a small base.

I've also updated some of the go packages, see the diff.

* update(Dockerfile): add bash

* update(Dockerfile): add missing packages

* update(Dockerfile): modify symlink

* add apk upgrade to fix alpine cves

Some new CVEs have popped up but adding apk update && apk upgrade to the Dockerfile ensures we are using the latest and most secure packages.

goofys (alpine 3.20.2)

Total: 2 (UNKNOWN: 0, LOW: 0, MEDIUM: 2, HIGH: 0, CRITICAL: 0)

┌────────────┬───────────────┬──────────┬────────┬───────────────────┬───────────────┬─────────────────────────────────────────────────────────┐
│  Library   │ Vulnerability │ Severity │ Status │ Installed Version │ Fixed Version │                          Title                          │
├────────────┼───────────────┼──────────┼────────┼───────────────────┼───────────────┼─────────────────────────────────────────────────────────┤
│ libcrypto3 │ CVE-2024-6119 │ MEDIUM   │ fixed  │ 3.3.1-r3          │ 3.3.2-r0      │ Issue summary: Applications performing certificate name │
│            │               │          │        │                   │               │ checks (e.g., ...                                       │
│            │               │          │        │                   │               │ https://avd.aquasec.com/nvd/cve-2024-6119               │
├────────────┤               │          │        │                   │               │                                                         │
│ libssl3    │               │          │        │                   │               │                                                         │
│            │               │          │        │                   │               │                                                         │
│            │               │          │        │                   │               │                                                         │
└────────────┴───────────────┴──────────┴────────┴───────────────────┴───────────────┴─────────────────────────────────────────────────────────┘

* update(Dockerfile): pin alpine image version to 3.20.3

I've also reverted the script to use bash instead of sh since we installed bash for compatibility reasons.

I've reverted the golang builder containers to use version 1.20.7 since we don't want to introduce new unknown issues.

---------

Co-authored-by: Bryan Paget <bryan.paget@statcan.gc.ca>